+++ CHANGELOG.md
@@ -14,6 +14,8 @@ way to update this template, but currently, we follow a pattern:
 
 ## Upcoming version 2025-XX-XX
 
+- [fix] SearchPageWithMap: secondary filter count was not taking into use filters that were limited
+  to category or listing type [#606](https://github.com/sharetribe/web-template/pull/606)
 - [fix] Handle multiple search page routes outside search page and clarify path param usage
   [#605](https://github.com/sharetribe/web-template/pull/605)
 
@@ -21,10 +23,8 @@ way to update this template, but currently, we follow a pattern:
 
 - [fix] Fix listing type path param usage
   [#599](https://github.com/sharetribe/web-template/pull/599)
-  
 - [fix] Avatar: use correct pending-approval variant link when user is pending approval.
   [#601](https://github.com/sharetribe/web-template/pull/601)
-  
 - [add] Add currently available translations for DE, ES, FR.
   [#597](https://github.com/sharetribe/web-template/pull/597)